Check Samples and Libraries for Correct Format in Pipeline (#1329)
Expands on using the formatter from the command-line.
 - Allows folders to be given.
- Adds the `--recursive` flag for specifying if you want the folders to
be processed recursively. Without this, if given a folder it will only
search the direct children of the folder.
- Adds the `--write` flag for specifying that the formatter should
overwrite unformatted Q# files with the format corrections. Without this
flag, the formatter only checks if the Q# files are correctly formatted.
To use the command-line formatter run `cargo run --binqsc_formatter --
<args to formatter>`.

Adds a check to the CI pipeline to run the formatter against the
libraries and samples, and stops a build if it finds files that are not
formatted correctly.