Initial integration of RCA, preview QIR gen, and adaptive profile (#1347)
In order to support adaptive profile selection in VS we have to add
`Adaptive` as a compilation attribute in the std lib. This is a
workaround until we design a full solution to address #996. This PR adds
adaptive as a profile.

Functionality to enable adaptive profile VSCode is hidden behind a
config value, as is using the preview QIR code generation from VSCode.

QIR generation for Adaptive/Unrestricted from Python will continue to
raise exceptions.

Adaptive is a profile allowed in the Python APIs, but warnings will be
issued to users indicating that this functionality is in preview.

The language service will surface RCA errors in VSCode and Python, but
not the Notebooks (will be handled in a follow up PR).
Other integration features are being tracked in #1294 and this PR is not
expected to have all functionality enabled.

---------

Co-authored-by: Alex Hansen <alex@alex-hansen.com>
Co-authored-by: Stefan J. Wernli <swernli@microsoft.com>
Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>