Add Q# code generation from AST. (#1380)
This PR uses the formatter and a new code gen function to turn a parsed
AST into the equivalent Q#. While the AST is a 1:1 representation of the
code, it isn't exactly what is in the code. We don't keep number
suffixes, scientific notation, readability separators, etc.

The test compile the sample Q#, generating new Q# from the AST and
getting a despanned AST package. It then feeds the new Q# back into the
compiler and repeats getting Q# and a despanned AST.

We verify that the Q# generated from the original and the Q# that had
been generated from the original, match. This verifies that the original
code and the output both generate the same output. This is not enough as
they may change the semantics.

We then compare the despanned AST from the original compilation to the
despanned AST from the generated Q#. This verifies that the semantics
from the original program were maintained through the round tripping.

While the former may seem unnecessary, it is very helpful in debugging
any failing tests as we get rich output of the expected Q# and expected
AST.

This introduces a dev-dependency to give very detailed string comparison
in tests which Rust's built-in `assert_eq` lacks. It would be great if
`expect-test` exposed their error diff implementation, but they don't.
It hasn't been updated in years, but appears feature complete and
stable. We can vendor if need be, but I couldn't find a good
alternative.