Update doc comments in std library (#1401)
This PR updates doc comments in the standard library. It also adds
information to the readme file describing what to put in different
sections of doc comments.
This is mostly random updates to a few spots in the library rather than
a consistent sweep over all functions and operation. Mostly to solicit
feedback on changes needed.

---------

Co-authored-by: Dmitry Vasilevsky <dmitryv@microsoft.com>
Co-authored-by: Scott Carda <55811729+ScottCarda-MS@users.noreply.github.com>
Co-authored-by: César Zaragoza Cortés <cesarzc@microsoft.com>