Avoid RCA panic when last stmt is binding
RCA had an assumption that a block with a non-Unit type must end with an expression, but as it turns out it's valid to have the outermost block of a callable with non-unit value end in an unreachable binding (a binding that appears after the return expresion, for example). If that code was submitted, the panic in RCA would cause a the language service to stop responding. This fixes the issue by treating the check for a terminating expression as best effort, since other logic already handles properly annotating callables that have return expressions.