Include return type of entry point in RCA check (#1417)
This change adds new runtime feature flags for use of specific types in
output from the entry expression/entry point that RCA can annotate a
program with. This allows the later check pass to verify that the code
will use types in output recording that are supported for the given
target. For example, a target that does not support dynamic integers
cannot accomodate recoding output with integer values, even if they are
static.