Update RCA handling of array dynamism (#1454)
This change updates the assumptions in RCA to be less restrictive on
assuming the dynamic size of arrays. Before, generating the `ValueKind`
for arrays would always assume both dynamic content and dynamic size,
which could cause failures in checks when, for example, a tuple of
static sized arrays is returned from the entry point. Instead, we rely
on the propagation of the runtime feature flags to track whether a
dynamically sized array is used. This is a precursor to later
simplifying the `QuantumProperties` struct by removing `ValueKind` in
favor of just tracking dynamic vs static.

---------

Co-authored-by: César Zaragoza Cortés <cesarzc@microsoft.com>