Fix lint message formatting (#1444)
When hovering over the squiggle generated by a Lint, the message in the
first line appears twice. This PR removes the duplicated message. Fixes
#1443


Currently, the error message is shown in the first line and in the last
line, which is redundant:

![image](https://github.com/microsoft/qsharp/assets/156957451/09edb462-b244-42e5-9831-d51da38ba284)

This is how it looks after the change:

![image](https://github.com/microsoft/qsharp/assets/156957451/ac9f5926-05b8-4e54-bd0f-e077d0a14186)