Use markdown-it and katex (#1438)
This addresses #845 

Previously we used 'marked' and 'mathjax' in the playground, but
'markdown-it' and 'katex' in VS Code - the latter being what VS Code
uses for its Markdown and LaTeX rendering in file previews, Jupyter
notebooks, etc. This converts the playground, and the kata content
generation, to use markdown-it and katex, and removes 'marked' and
'mathjax' from the repo.

Note that KaTeX is more strict that MathJax that the LaTeX is well
formed. This is probably a _good thing_ while we're authoring content. I
did need to clean up quite a bit of LaTeX in testing this, and there's
probably quite a bit more elsewhere to go over. (In general our LaTeX is
often in poor shape. We might want to look into any 'linting' options
for it). Feel free to skip the .md files and focus on the code
initially.

I also significantly revamped how we render markdown in a Preact
components. This has led to much cleaner code, and deleting a lot of
useRef and useEffect and dangerouslySetInnerHtml code, and the need to
re-typeset content. (Esp. see kata.tsx and docs.tsx).

---------

Co-authored-by: Mariia Mykhailova <mamykhai@microsoft.com>