Handle impossible factories in RE API (#1463)
Currently, the factory builder API returns a vector of factories for
space-time evaluation. However, returning no factory may mean that the
algorithm was too fast, and the optimizer will continue to search for
some. But there are some cases in which no factory can be found given
the input parameters, no matter of the duration of the algorithm.

This PR changes the API to allow to indicate if no magic factories can
be computed by returning `Option<Vec<...>>` instead of `Vec<...>`.