Fix global phase for `PauliI` rotation and `DumpRegister` (#1461)
This change introduces a new, stdlib-only intrinsic for applying global
phase during simulation to make phase behavior more consisten and fixes
#1450. This intrinsic only applies the global phase during simulation;
the new `GlobalPhase` intrinsic is a no-op for QIR generation, circuit
generation, and resource estimation. The change also includes updates to
the stdlib and associated tests to account for the corrected global
phase behavior.

This also fixes a bug in `DumpRegister` related to iteration through the
state vector that ensures the "base" amplitude used for separating the
state is consistently chosen, which avoids corner cases that could
introduce a phase in the displayed output that wasn't present in the
underlying state.