LaTeX processing improvements (#1475)
While I had the context paged in, I threw this together to make life
easier going forward.

With the prior solution, there were some problems that had some
'creative' workaround that would have been challenging for others going
forward to catch or remember. This avoids need to use those. (See
comments in PR for details).

Most of the code is lifted from the exiting markdown-it katex extension,
so is well tested. (And I've tested thoroughly myself and ensured output
is consistent).