Fix use of `ForEach` in samples for compatibility (#1520)
The Bernstein-Vazirani and Hidden Shift samples both used a pattern that
is not well supported by RCA, specifically calling `ForEach(MResetZ,
qubits)` to perform measurement on each qubit in an array. Passing a
callable that produces a dynamic value makes static analysis very
challenging, so instead we can update these to use the equivalent
`MResetEachZ` operation. This allows the samples to work well with
analysis and produce QIR for both base and adaptive profile.