Fix debugger scope handling during early returns (#1528)
Changes introduced in #1442 to enable debugger tracking of variable
scopes caused a bug if a program used callables with early returns,
namely those returns would not pop any added scopes and later execution
would read/write the wrong variable states. This fixes the issue by
introducing a debug-only return node that triggers a frame leave
handling in the environment. This needs to be debug only because it
should not be used during partial evaluation.