Allow implicit top-level namespaces when parsing files (#1416)
This PR allows for implicit top-level namespaces.

Prior to this PR, all Q# had to be wrapped in namespace declarations:

```qsharp
namespace Foo {
  operation Main() : Unit {}
}
```

Now, users can optionally omit the top-level namespace declaration. This
is nice, because it removes a level of necessary indentation.

If the explicit namespace declaration is omitted, the file path is used
to determine the namespace name. For example, in
`~/code/qsharp/project_dir/src/Foo/Bar/Baz.qs`, the file:
```qsharp
operation Quux() : Unit {}
```

defines a callable in the namespace `Foo.Bar.Baz`.

We detect the "root" of the project by calculating the largest common
prefix of the file paths. This is the most versatile approach, in my
opinion, without having to introduce project-based logic into the
compiler itself (that is, add a `qsc_project` dependency to
`qsc_frontend` -- which would be cyclical via `qsc_lint`). It allows us
to have a perfect UX for the happy path following a canonical project
layout, and reasonable UX for other situations (e.g., manually evoking
`qsc` on two files in different parts of your computer -- what is the
"root"?).

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>
Co-authored-by: Mine Starks <16928427+minestarks@users.noreply.github.com>