Test samples with and without debugging
This change introduces a scenario test under qsc that verifies the behavior of the samples with normal and debug execution. It uses fixed seeds to ensure results are predictable and restricts the slowest test (Shors.qs) to be release only.

Included is some cleanup of output from the samples so that compatibel characters are used and no trailing spaces are present.