Support target name in Python, remove Adaptive warnings
This change adds an optional `target_name` property to `qsharp.init()` where the user can specify the target and have the profile inferred. Right now it just uses some hard-coded logic to decide between "ionq", "rigetti", or "quantinuum" and anything else is treated as unrecognized with a prompt for the user to specify `target_profile` directly.