Surface partial evaluation errors in VS Code (#1552)
- Defines a struct `IQSharpError` at the WASM boundary to contain full
error information from Q# interpreter error calls. Currently this is
only used in QIR generation and circuit visualization, but it can and
should eventually be wired up for all `ICompiler` methods that can
return errors.

- We create a new diagnostic collection (a set of errors to be shown in
the Problems view in VS Code) that shows errors from the last run
command. These errors persist in the editor, but can be dismissed by the
user or when another command is run.


![partial-eval-errors](https://github.com/microsoft/qsharp/assets/16928427/1a54866c-8e45-475d-8cdb-e5b10cee5b45)