Updated UX for VSCode QIR generation (#1548)
This change updates the extension behavior for QIR code generation to
give more granular error messages and guidance when possible. We used to
always report an error with "Submitting to Azure" and recommend
switching to Base profile. Now the message and guidance will
differentiate between a few different scenarios:
- Generating QIR with profile set to unrestricted will use codegen
specific message and prompt for Base Profile.
- Submitting to Azure target with profile set to unrestricted will
recommend either Base or Adaptive_RI depending on what the target
supports.
- Submitting to Azure target that only supports Base profile with
Adaptive_RI configured will give specific error message and recommend
Base.

This also fixes a minor bug with a missing await that could cause an
error message to crop up even if the user clicked "Set" to take the
recommended change.