Drop FIR increment if FIR passes fail
This fixes an issue with incremental compilation where an increment with a failure could still have that FIR present in the package during later compilation, making errors "sticky" instead of associated with just the increment. This achieves it by tracking what the most recently lowered additions to the package were and removing them if there were pass failures like those from RCA check.