Update RCA for mutable `Result`, array, and tuple variables
This change strengthens RCA checks around use of mutable `Result`, array, and tuple variables in a dynamic context to make them consistent with how we verify the use of other dynamic value types like `Int` and `Double`. This gives users compile time feedback when updating a mutable value in a dynamic context will not produce QIR, rather than a codegen failure with `an unexpected error occurred related to: re-assignment within a dynamic branch is unsupported for type <type>` error.