Fix panic when updating array with dynamic value (#1606)
This fixes a panic that was introduced in #1571, where if the content
being used to update an array was a dynamic, non-array value it would
panic when trying to aggreate a default array type into a element type.
To make sure the update matches the chosen type, generate the
`ValueKind` from the matching expression type. Added unit tests for both
array and non-array dynamic update types.