[Unitary Hack] Save Python qsharp-widgets RE output to PNG (#1604)
Submission for Unitary Hack issue: fixes #1063 

This change adds the ability to save the resource estimates diagram
locally as a PNG. For this I added a new dependency called `html2canvas`
that takes a DOM and renders it to an image. This currently works only
on IPython `qsharp-widgets`, could be extended to the VSCode version.

Additionally we take into account the color theme within VSCode (which
is provided in the stylesheet `qsharp-ux.css`) and this is passed in as
the background color for rendering. Tested on (dark/light) Modern VSCode
themes.



https://github.com/microsoft/qsharp/assets/3744263/898de6d9-fba4-4fea-b729-a3bb698dc51c



~Current issues:~
- ~This button is placeholder, nicer version:
https://github.com/microsoft/vscode-notebook-renderers/pull/118~
- ~Saved imaged is slightly truncated at the bottom~

---------

Co-authored-by: Mine Starks <mineyalc@microsoft.com>