Refactor: Consolidate "active Q# program" logic in VS Code (#1623)
This is in preparation for the package work, but I think it's a useful
refactor in its own right. No behavior changes.

There are quite a few features/commands across the VS Code extension*
that are meant to operate on the "currently active Q# project". This is
defined as the Q# document in the currently active editor, and if that
document is associated with a qsharp.json, the entire project.

This code was duplicated across all these commands, and was slightly
different everywhere. There have been bugs where we weren't respecting
some of the common settings (e.g. #1565) and changes where we had to
update each command separately if we changed something about project
semantics (e.g. #1089, #918) . This PR aims to unify all that so we
touch fewer places when we change something about the project compiler
configuration logic.

*Run, Debug, Circuit, Estimate, Histogram commands, Show Documentation,
Get QIR and Submit to Azure.