Show code lenses on Main() even when @EntryPoint() attribute is omitted (#1652)
With #1584 , the `@EntryPoint()` attribute is no longer required if the
entry callable is named `Main` . But the code lenses are still looking
for the attribute.

With this fix, instead of looking for the `@EntryPoint()` attribute, we
parse the entry expression. This expression would have been generated by
a prior HIR pass.
https://github.com/microsoft/qsharp/blob/54586a6bb9d91f21cbed1cbc5359a7a8ae43db64/compiler/qsc_passes/src/entry_point.rs#L44