Option to load custom RE parameters in DF chemistry sample (#1662)
This adds a program argument `-p` (long `--paramsfile`) to the Python
script of the DF chemistry sample to have the option to load one or
multiple RE configurations that are stored as objects or arrays in JSON
files. If no such argument is specified, the default configuration is
used as before.

This addresses the second bullet point in #1115.