Parameter to set maximum number of distillation rounds in default search (#1669)
This adds a parameter `maxDistillationRounds` to the `constraints` group
in resource estimation, which allows to modify the allowed number of
distillation rounds in the first search for T factories. The default
value is 3. Making this number larger will slow down resource estimation
due to the larger exploration space of possible T factories.