Lints and Code Actions for Deprecated UDT Syntax (#1660)
Expands the linting framework and adds lints for `newtype` declarations,
references to UDT constructor functions, use of the `::` field accessor
operator, and use of the `w/` and `w/=` operators for structs. A code
action is added to convert uses of `w/` and `w/=` operators with structs
to copy constructor syntax, and another code action is added to convert
uses of `::` to `.`.

---------

Co-authored-by: Oscar Puente <156957451+orpuente-MS@users.noreply.github.com>