Skip commenting memory/perf benches on draft PRs (#1694)
This should help reduce some of the comment noise until we have time to
customize that action some more.

I tested that this works on this PR -- the benches are not run when the
PR is a draft.

I had to pull the two relevant jobs out into a different workflow file,
because the triggers are different. For the reporting benches stage (the
memory and perf comments that show up on PRs), they'll only run if the
PR is in a draft. But, and this is the different part that required them
being in a separate file, they also need to trigger when a PR is marked
"ready for review".