Q# Package System (#1698)
This PR:

1. Refactors the compiler API such that the caller passes in a
constructed package store, allowing for dependencies to be pre-compiled
and inserted before compiling user code
2. Introduces the `BuildableProgram` abstraction, which encapsulates the
notion of "some user code that is ready to compile, along with a package
store that has all of its dependencies compiled"
3. As a drive-by fix, fixes circuits for internal operations
4. Removes the notion of `Visibility` from the AST, since visibility is
calculated via exports now. `internal` is still parsed without error for
backwards compatibility, but is a no-op because `internal` is the
default now.
5. Enforces visibility in the HIR, where items must be
`Visibility::Public` to show up across packages as a `GlobalItem`
6. Adds export statements to the standard library to preserve the
existing API with the new internal-by-default semantics
7. Inserts all packages into the user code's namespace tree with the
prefix defined as the package alias in the `qsharp.json`



closes #883 


The Q# formatting CI stage will fail until #1692 goes in

This PR supersedes #1693