Enable Run menu in VS Code
The Run menu options for "Start Debugging" and "Run without Debugging" were not working, in part to how the validation checking for the presence of a launch.json. This reverses the if-statement and drops the else condition to get recognition of single file scenarios without launch.json. Confirmed via manual testing.