In libraries, treat a top-level namespace called Main as the root of the library (#1705)
Currently, if you have a library with the following structure:

```
src/
   Foo.qs
   Bar.qs
```

And you import it:

```
{
  "dependencies": { "MyLibrary": ... } 
}
```


The API will always include at least two initial path items: the package
alias (`MyLibrary`) and the namespace (`Foo` or `Bar`). Every exported
item will be `MyLibrary.Foo.X()` or `MyLibrary.Bar.X()`. There's no way
to construct
a library export such that it is referenced at the top level, e.g.
`MyLibrary.X()`.

This PR solves this problem. In a library, if there's a namespace called
`Main`, it is treated as the root of the library. So, if you have the
below structure:

```
src/
   Foo.qs
   Main.qs
   Bar.qs
```

Anything exported from `Main.qs`, say a callable `X`, will show up in
the consumer of the library as `MyLibrary.X()`. The behavior of items
exported from `Foo.qs` and `Bar.qs` is unchanged.