Languages Service loads project and dependencies in Notebooks (#1717)
This updates the compilation logic for the language service when the
active document is a notebook to include any specificed project root
folder such that dependencies are correctly resolved. This avoids
spurious error squiggles on items that come from dependencies and even
enables go to definition and hover for those items.