Use `PackageType::Lib` by default
In most places, we already used the `PackageType::Lib` by default, but not in the language service. With more library authoring expected in the future, this tweaks the user experience to avoid errors for missing entry point until the Q# program as run. As part of these chagnes, the behavior of the entry point check is updated to always provide feedback for other entry point related errors (like duplicates) but only return an error on missing entry point if the compilation is for an executable.

Of note, this does NOT remove the configuration option for `packageType` in the qsharp.json, which we can do as a follow up if needed.