Use `PackageType::Lib` by default (#1737)
In most places, we already used the `PackageType::Lib` by default, but
not in the language service. With more library authoring expected in the
future, this tweaks the user experience to avoid errors for missing
entry point until the Q# program as run. As part of these changes, the
behavior of the entry point check is updated to always provide feedback
for other entry point related errors (like duplicates) but only return
an error on missing entry point if the compilation is for an executable.

Of note, this does NOT remove the configuration option for `packageType`
in the qsharp.json, which we can do as a follow up if needed.

This does update the error string that pops up when failing to run/debug
the Q#. If the program is missing an entry point, a more specific error
message is returned:

![image](https://github.com/user-attachments/assets/1e667ac3-dd3d-4c04-b393-49c8ceb1097d)

and the generic compilation failure message is also updated:

![image](https://github.com/user-attachments/assets/78fa1b8c-5ad9-4806-b501-28e01a546049)